Bring blkdat.input up to date
The blkdat variables that have been added or changed are: hybthn, optional dx0k, mtracr, and epmass=2 for river only mass exchange.

hybthn is HYCOM-src commit 8d24702
dx0k is HYCOM-src commit 4334a88
mtracr is HYCOM-src commit 4982038
epmass=2 is HYCOM-src commit 7661e14

hybthn is set to 1.0 to recoveer the original behaviour, but 5.0 is more typical.
dx0k is not present here, and mtracr=0 disables diagnostic tracers.  The original
had empbal=1 and epmass=1, which has been changed to empbal=0 and epmass=2.