Bring blkdat.input up to date, II
The blkdat variables that have been added or changed are: hybthk, mstrcr,
strflg, tidfXX, drgscX and ocnscl.

hybthk is HYCOM-src commit 5fb66b6
mstrcr is HYCOM-src commit e01c258
tidfXX is HYCOM-src commit f32b690
ocnscl is HYCOM-src commit d3f5330

hybthk is set to 0.0 to recoveer the original behaviour, try 0.9.
mstrcr and strflg are setup to add oneta to the archs files.
If mstrcr=0 then there are no strflg entries.
tidfXX is setup for no streaming filters.
ocnscl is 0.0 to confirm the absolute winds from wndflg=-4.

In addition, SHASTA has been added as an OS option for the scripts.
This is a HPE Cray EX system with 128 AMD cores per node.